requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.43
python-multipart==0.0.20
pandas==2.3.2
openpyxl==3.1.5
xlrd==2.0.2
pyxlsb==1.0.10
python-docx==1.2.0
extract-msg==0.55.0
pywin32==311
pdfplumber==0.11.7
pypdf==6.0.0
pypdfium2>=4.30.0
pillow==11.3.0
opencv-python==4.12.0.88
pytesseract==0.3.13
streamlit==1.49.1
pytest==8.4.1
httpx==0.28.1
numpy==2.2.6
pyyaml>=6.0
rapidfuzz>=3.9.0
img2pdf>=0.5.1
alembic>=1.13.0
